Tasarım yapıldı

+++ MMEntityFrameworkMarket/Product.cs
@@ -11,7 +11,7 @@ namespace MMEntityFrameworkMarket
         public int Id { get; set; }
         public string Name { get; set; }
         public decimal Price { get; set; }
-        public int StockAmount { get; set; }
+        public decimal StockAmount { get; set; }
         public string StockAmountType { get; set; }
         public string Category { get; set; }
     }
